Relocated to test_join.py
